명세 변경 반영 및 문제 중복 방지

+++ backend/src/main/java/com/aimong/backend/domain/auth/entity/ChildProfile.java
@@ -86,7 +86,7 @@ public class ChildProfile {
     private ProfileImageType profileImageType;
 
     @Column(name = "session_version", nullable = false)
-    private int sessionVersion;
+    private long sessionVersion;
 
     @Column(name = "fcm_token")
     private String fcmToken;
